backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "awdio-backend"
version = "0.1.0"
description = "Voice-driven podcast platform backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",

    # Settings & validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # AI & ML
    "openai>=1.57.0",
    "pyneuphonic>=1.0.0",

    # Document processing
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",

    # Object storage
    "minio>=7.2.0",

    # WebSockets (pinned for pyneuphonic compatibility)
    "websockets>=12.0,<13.0",

    # Utils (pinned for pyneuphonic compatibility)
    "httpx>=0.27.2,<0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
